# feat(jvm): Optimize JVM  GC logging for arm64