--- a/platform/src/main/java/common/org/tron/common/arch/Arch.java
+++ b/platform/src/main/java/common/org/tron/common/arch/Arch.java
@@ -69,11 +69,15 @@ public final class Arch {
     return javaSpecificationVersion().equals("1.8");
   }
 
+  public static boolean isJava17() {
+    return javaSpecificationVersion().equals("17");
+  }
+
   public static void throwIfUnsupportedJavaVersion() {
-    if (isX86() && !isJava8()) {
+    if ((isX86() && !isJava8()) || (isArm64() && !isJava17())) {
       logger.info(withAll());
       throw new UnsupportedOperationException(String.format(
-          "Java %s is required for %s architecture. Detected version %s", "1.8",
+          "Java %s is required for %s architecture. Detected version %s", isX86() ? "1.8" : "17",
           getOsArch(), javaSpecificationVersion()));
     }
   }